finished scaffolding for all files - animat now shown on screen walking. although i think it might be walking upside down?
