+++ ka/hinged_segment.py
@@ -1,66 +1,35 @@
 from __future__ import annotations
+import matplotlib.pyplot as plt
 
-# from hinged_segment import HingedSegment  # stops the errors, remove later
-from ka.point import Pt
+from unittest.mock import Mock
 
 from typing import List, Tuple, Optional, Union
-
+from icecream import ic
 
 from math import atan2, cos, degrees, pi, sin, sqrt
 
+from point import Pt
+from gaits import Gait, FootState
 
-class HingedSegment:
-    def __init__(
-        self,
-        global_angle: float,
-        length: float,
-        parent_or_location: Union[HingedSegment, Pt],
-        max_angle: float = 90.0,
-    ) -> None:
-        """Takes a global angle, decides if it has a parent, then creates a HS object.
 
-        Args:
-            global_angle (float): angle in relation to y axis and ground
-            length (float): length of segment
-            parent_or_location (Union[HingedSegment, Pt]): either contains a parent HS object, or a starting
-                location if no parent is given
-        """
-        # Angle is always with respect to the global x-y coordinate system
-        self.ang = global_angle
-        self.len = length
+class HingedSegment(object):
+    def __init__(self, start: Pt, angle: float, len: float):
+        # self.start = start
+        self.start = Pt(0, 0)
+        self.angle = angle
+        self.len = len
 
-        # For type checking
-        self.loc = Pt(0, 0)
-        self.par: Optional[HingedSegment]
-        self.chi: Optional[HingedSegment] = None
-        self.hip = None
+        self.end = self.calculate_end()
 
-        if isinstance(parent_or_location, HingedSegment):
-            # if what was passed in is a HS, it is a parent
-            self.par = parent_or_location
-            self.update_from_parent(0)
-        else:
-            # here, the parent_or_location var doesn't have type HS, so it can't be a parent.
-            #  it contains the location of a hip
-            self.par = None
-            self.hip = parent_or_location
-            # self.loc = parent_or_location
+        self.chi = None
 
-    def __str__(self) -> str:
-        """Represents object as string containing info on location and angle
+    def calculate_end(self) -> Pt:
+        x = self.start.x + self.len * cos(self.angle)
+        y = self.start.y + self.len * sin(self.angle)
+        return Pt(x, y)
 
-        Returns:
-            str: str(self.loc) + f" @ {degrees(self.ang): .3f}°"
-        """
-        return str(self.loc) + f" @ {degrees(self.ang): .3f}°"
-
-    def __repr__(self) -> str:
-        """returns same thing as __str__, for times when str() is not automatically called
-
-        Returns:
-            str: [description]
-        """
-        return self.__str__()
+    def add_child(self, chi: HingedSegment):
+        self.chi = chi
 
     def update_from_parent(self, delta_angle: float) -> None:
         """propagates changes in angle from parent HS to child HS objects
@@ -68,9 +37,8 @@ class HingedSegment:
         Args:
             delta_angle (float): amount by which angle will change
         """
-        self.ang += delta_angle
-        self.loc.x = self.par.loc.x + self.par.len * cos(self.par.ang)
-        self.loc.y = self.par.loc.y + self.par.len * sin(self.par.ang)
+        self.angle += delta_angle
+        self.end = self.calculate_end()
         if self.chi:
             # update child if present
             self.chi.update_from_parent(delta_angle)
@@ -81,27 +49,13 @@ class HingedSegment:
         Args:
             new_angle (float): new angle for this segment
         """
-        delta_angle = new_angle - self.ang  # the difference between new and old angles
-        self.ang = new_angle
+        delta_angle = (
+            new_angle - self.angle
+        )  # the difference between new and old angles
+        self.angle = new_angle
+        self.end = self.calculate_end()
         if self.chi:
             self.chi.update_from_parent(delta_angle)
 
-    def get_tip_location(self) -> Pt:
-        """Gets the x and y location of the tip of each segment
-
-        Returns:
-            Pt: position of tip
-        """
-        tip = Pt(0, 0)
-        tip.x = self.loc.x + self.len * cos(self.ang)
-        tip.y = self.loc.y + self.len * sin(self.ang)
-        return tip
-
-    def get_hip(self):
-        if self.hip == None:
-            return self.par.get_hip()
-        else:
-            return self.hip
-
-    def add_child(self, chi: HingedSegment):
-        self.chi = chi
+    def get_tip(self):
+        return self.end
